# Removing BacktestingTransactionHandler thread