--- a/Common/Util/BusyBlockingCollection.cs
+++ b/Common/Util/BusyBlockingCollection.cs
@@ -1,11 +1,11 @@
 ﻿/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,6 +18,7 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Util
 {
@@ -26,7 +27,7 @@ namespace QuantConnect.Util
     /// being processed
     /// </summary>
     /// <typeparam name="T">The item type being processed</typeparam>
-    public class BusyBlockingCollection<T> : IDisposable
+    public class BusyBlockingCollection<T> : IBusyCollection<T>
     {
         private readonly BlockingCollection<T> _collection;
         private readonly ManualResetEventSlim _processingCompletedEvent;