--- a/Engine/TransactionHandlers/BacktestingTransactionHandler.cs
+++ b/Engine/TransactionHandlers/BacktestingTransactionHandler.cs
@@ -1,11 +1,11 @@
 /*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,6 +17,8 @@ using System;
 using QuantConnect.Brokerages.Backtesting;
 using QuantConnect.Interfaces;
 using QuantConnect.Lean.Engine.Results;
+using QuantConnect.Orders;
+using QuantConnect.Util;
 
 namespace QuantConnect.Lean.Engine.TransactionHandlers
 {
@@ -40,10 +42,13 @@ namespace QuantConnect.Lean.Engine.TransactionHandlers
             {
                 throw new ArgumentException("Brokerage must be of type BacktestingBrokerage for use wth the BacktestingTransactionHandler");
             }
-            
+
             _brokerage = (BacktestingBrokerage) brokerage;
 
             base.Initialize(algorithm, brokerage, resultHandler);
+
+            // non blocking implementation
+            _orderRequestQueue = new BusyCollection<OrderRequest>();
         }
 
         /// <summary>
@@ -51,6 +56,9 @@ namespace QuantConnect.Lean.Engine.TransactionHandlers
         /// </summary>
         public override void ProcessSynchronousEvents()
         {
+            // we process pending order requests our selves
+            Run();
+
             base.ProcessSynchronousEvents();
 
             _brokerage.SimulateMarket();
@@ -66,6 +74,32 @@ namespace QuantConnect.Lean.Engine.TransactionHandlers
 
             _brokerage.SimulateMarket();
             _brokerage.Scan();
+        }
+
+        /// <summary>
+        /// For backtesting we will process the order ourselves
+        /// </summary>
+        /// <param name="ticket">The <see cref="OrderTicket"/> expecting to be processed</param>
+        protected override void WaitForOrderToBeProcess(OrderTicket ticket)
+        {
+            // we process the order request our selves
+            Run();
+
+            if (!ticket.OrderSet.WaitOne(0))
+            {
+                // this shouldn't happen, but just in case...
+                throw new Exception("BacktestingTransactionHandler.ProcessRequest(): " +
+                    $"The order request (Id={ticket.OrderId}) was not processed.");
+            }
+        }
+
+        /// <summary>
+        /// For backtesting order requests will be processed by the algorithm thread
+        /// sequentially at <see cref="WaitForOrderToBeProcess"/> and <see cref="ProcessSynchronousEvents"/>
+        /// </summary>
+        protected override void InitializeTransactionThread()
+        {
+            // nop
         }
     }
 }