… Tests/Algorithm/AlgorithmLiveTradingTests.cs
… Tests/Common/Orders/Fills/PartialMarketFillModelTests.cs
… Tests/Common/Securities/CashBuyingPowerModelTests.cs
… Tests/Engine/BrokerageTransactionHandlerTests/BacktestingTransactionHandlerTests.cs
… Tests/Engine/BrokerageTransactionHandlerTests/BrokerageTransactionHandlerTests.cs
… Tests/QuantConnect.Tests.csproj